doctor_syria/requirements_new.txt
# AI and Machine Learning
tensorflow>=2.8.0
scikit-learn>=1.0.2
opencv-python>=4.5.5.64
pillow>=9.0.1

# IoT and Real-time
paho-mqtt>=1.6.1
channels>=3.0.4
channels-redis>=3.4.0

# Medical Imaging
pydicom>=2.3.0
nibabel>=3.2.2

# API and Web
djangorestframework>=3.13.1
django-cors-headers>=3.11.0
django-filter>=21.1

# Database
psycopg2-binary>=2.9.3
redis>=4.2.2

# Task Queue
celery>=5.2.3
flower>=1.0.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Testing
pytest>=7.1.1
pytest-django>=4.5.2
factory-boy>=3.2.1

# Development
black>=22.3.0
flake8>=4.0.1
mypy>=0.942
